Add image_input library to read still images
Add a CMake build with a subdirectory containing the image_input
library. This sub-project wraps the use of the IO extension for
Boost::GIL, which lets us compile the rest of the project with
C++20.

Adds a demo application which reads a JPEG file and a PNG file
and prints some metadata about them.