Add image writing to img lib  and rename to image_io
Adds two functions to the image library to write jpeg and png
images to disk using the GIL IO extension. Renames the library
to image_io from image_input since it now has output capabilities.

The writing functionality won't be used as part of the final
product, but it will be useful for debugging any issues with
the colour table/palletization without relying on the
correctness of other modules like the LZW encoder or the GIF
data stream packer.