Implement lzw::lzw_encoder::encode and flush
Fully implement the LZW encoder and add tests for most of its
functionality. The only major missing test case is to verify the
encoder's behaviour when the dictionary reaches the 4096-entry
limit imposed by the specification.
It may also be worth revisiting the dictionary implementation to
use a more memory-efficient structure like a prefix tree instead
of a std::map.