Add a palettization module and interfaces for it
Add some typedefs for image processing in common/image_utils.hpp.

Add a palettize subproject that provides interfaces for generating
a color table for an image and using that color table to quantize
an image into a set of indices which are lzw-encodable.

This subproject will be built out with placeholder implementations
to start, and a second pass will be done later in the development
process to implement intelligent palettization with the  Median Cut
algorithm.