Move GIF block construction to gif package
Removes the GIF data formatting logic from gifgen.cpp and into
a builder class that will more cleanly support further extensions
(graphics and netscape application blocks and multiple images).

There is an issue with the includes for the gif package as it
depends on some of its sibling packages for LZW and palettization
utilities. Relative paths are used for now, but should be removed
once we know how to fix the issue with CMake.