Re-use color histogram for median cut region splitting
The Springer book's description of the algorithm makes is clear
that they first compute a color histogram, and then re-build a
list of all the colors with their frequencies into a similar
structure for use when splitting the regions. This commit removes
the first representation and constructs the histogram in a format
that can be used for splitting the regions as well, removing
O(# colors in image) work per-frame.