Remove relative paths in gif_builder includes
Removes relative include paths for palettize and lzw headers in
gif_builder.hpp/cpp. The LZW library is header only, and I wasn't
able to find a nice way to export the include directory directly,
so I added it as a global include. The palettize one is
automatically included in code that links against the static library.