Replace map of vectors with hash-map of strings in LZW encoder
Instead of using a vector of uint8_t's, a std::string is used
to hold the working symbol string. The underlying std::map
used for the LZW dictionary is replaced with an unordered_map.
The string-replacement on its own had a slight performance
improvement accross a variety of inputs, and the pair of
changes together resulted in about a 10% speedup of the entire
application.