Remove use of fmt::format from degif.cpp
The degif utility is the only part of the build that requires
fmt::format. Since the formatting library isn't yet avaliable in
the standard library and setting up the fmt library for the docker
build is a pain, we can just remove the formatted messages.

The degif utility was originally just for debugging my progress during
the course, so it isn't really needed anymore, but I'll keep it around
in case its useful for later features or for others working on similar
projects.