Add Dockerfile to build gifgen
Adds a Docker build file so we can build and run the gifgen application
from anywhere (that has docker installed). Most dependencies are the
same as those that were used for the original development, but some
versions are slightly different (minor versions +/-2).

Contraint on cmake version in CMakeLists.txt is relaxed so we can use
the version available via apt-get instead of building it ourselves.